Add CI/CD pipeline for Terraform plan/apply
Adds a simple CI/CD workflow for planning and applying TF changes.

Plan is run on pull requests and on pushes to main, whereas Apply is
only run on pushes to main. This means that a merge is an implicit
'approve this change to go live' action.

Having a dedicated approval process for deployment without Github Pro is
annoying.